A bunch of commits to reduce the size of the container (#3)
* Stop installing pip from pip

The UBI9 image is recent enough, that we can just rely on the
system-packaged pip. Should the need arise we can always backtrack
on that, but for now let's stick to the pip rpm we're being provided.

* Remove some unneeded collections

Namely we remove "fortinet cisco dellemc f5networks junipernetworks mellanox netapp"

This saves us ~400MB
Before:
localhost/utility-container                  latest      4cd658d42edf  28 seconds ago  1.97 GB

After:
localhost/utility-container                  latest      685c976bab41  22 seconds ago  1.54 GB

* Run pip install in a single command

No need now to have multiple invocations.

* Reduce the Azure sdk size impact

See https://github.com/Azure/azure-sdk-for-python/issues/11149
and https://github.com/Azure/azure-sdk-for-python/issues/17801
The size of the azure sdk is ridiculous because they keep old (and unused
APIs) around. We run this azure_sdk_trim.py to prune the unused APIs while
still maintaining a good compatibility level.

This shaves off about 0.5GB from the image. The full images is now
around ~1GB.

* Install git-core only

This avoids us installing perl-libs. The fundamental git commands
(clone) are in git-core and those are the only ones we are likely going
to need for the foreseeable future.

* Make kubectl a symlink to oc

This won't have an impact on the final size of the image or on du
calculations. But it makes the intent clearer as the two commands have
the exact same content.

* Add ARG variabls to remove some rpms

We have some rpms we remove via microdnf (fonts) and some
that we forcefully remove as we'll likely never ever need those
via rpm -e --nodepts (cracklib-dict)

* Move the openshift_client version to a variable

So it will be easier to change and there is less clustter in the
Containerfile

* Fix some small hadolint warnings

Quote any >= sign and move the -y as a first argument

* Fix package names indentation in README and add vi